example is much faster. Who knew, the loop I wanted to add a C file just to implement already exists right inside of simavr!
One concern I still have is that there's a lot of conversion to/from
absolute and relative terms in gooptest.lisp, with regards to
cycles-rel, cycles-abs, etc. Because the default implementation of (core-many-cycles) converts its relative argument back to absolute. So (cycles-abs) converts absolute to relative, core-many-cycles converts it back to absolute! But it is useful to have a relative argument to core-many-cycles because of, say, our specific avr implementation.